+++ blueprints/http-proxy/files/server/proxies/__name__.js
@@ -4,11 +4,14 @@ var Proxy = require('http-proxy');
 // https://github.com/nodejitsu/node-http-proxy
 var proxy = Proxy.createProxyServer({});
 
+var proxyPath = '/<%=camelizedModuleName %>';
+
 module.exports = function(app) {
 
-  app.use('/<%=camelizedModuleName %>', function(req, res, next){
+  app.use(proxyPath, function(req, res, next){
+    // include root path in proxied request
+    req.url = path.join(proxyPath, req.url);
     proxy.web(req, res, { target: '<%=proxyUrl %>' });
   })
 
 };
-